requirement.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
idna==3.10
pycparser==2.22
pycryptodome==3.21.0
PyKCS11==1.5.17
requests==2.32.3
urllib3==2.2.3
